requirements.txt
streamlit>=1.32.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
python-dotenv>=1.0.0
google-generativeai>=0.3.1
ffmpeg-python>=0.2.0